requirements.txt
anyio==3.7.0
attrs==23.1.0
autoflake==2.1.1
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
openapi-client==1.1.7
openapi-generator-cli==4.3.1
openapi-python-client==0.14.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
pluggy==1.0.0
pydantic==1.10.9
pyflakes==3.0.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
typer==0.9.0
typing_extensions==4.6.3
urllib3==2.0.2
